The position of a selection is now a size_t
It is more in line with good D practices. Also some :1vG= and compile
fixes.

+++ source/graphics/manipulation.d
@@ -130,7 +130,7 @@ void setTitle(Text title, string text)
 }
 
 
-void changeOpacity(ref ubyte[] opacities, const int position)
+void changeOpacity(ref ubyte[] opacities, const size_t position)
 {
     /*
       Change the opacity of background images for a more or less fluid transition.
